Changes for x64 port.
git-svn-id: http://svn.code.sf.net/p/spimsimulator/code@423 663c0655-963b-4438-a250-40b5657b8e18

+++ PCSpim/Regman.cpp
@@ -91,7 +91,7 @@ LONG RegMan_SetSZValue(HKEY hKey, LPCTSTR lpstrValueName, LPCTSTR lpstrSrc)
 		       0,
 		       REG_SZ,
 		       (LPBYTE)lpstrSrc,
-		       lstrlen(lpstrSrc) + 1);
+		       (DWORD)lstrlen(lpstrSrc) + 1);
 }
